third_party/blink/web_tests/fast/dom/HTMLAnchorElement/update-href-of-focused-anchor.html

<!DOCTYPE html>
<body>
<a href="javascript:" id="target">link</a>
<script src="../../../resources/js-test.js"></script>
<script>
jsTestIsAsync = true;
var target = document.getElementById('target');
window.onload = async function() {
    debug('Remove href.');
    target.focus();
    shouldBe('document.activeElement', 'target');
    target.removeAttribute('href');
    await new Promise(setTimeout);
    shouldBe('document.activeElement', 'document.body');

    debug('Update href.');
    target.href = 'javascript:';
    target.focus();
    target.href = 'javascript:undefined';
    shouldBe('document.activeElement', 'target');

    target.remove();
    finishJSTest();
};
</script>
</body>
