third_party/blink/web_tests/platform/generic/fast/dom/HTMLAnchorElement/update-href-of-focused-anchor-expected.txt

Remove href.
PASS document.activeElement is target
PASS document.activeElement is document.body
Update href.
PASS document.activeElement is target
PASS successfullyParsed is true

TEST COMPLETE
